requirements.txt
dotenv-python
groq
numpy
pandas
openai
langchain
langchain-community
pytubefix
streamlit